// src/utils/error/types.ts
export interface CustomErrorType extends Error {
  errors(errors: any): unknown;
  statusCode?: number;
  code: number;
}

type error = {
  message: string;
  errorDetails?: CustomErrorType;
};

export interface ErrorResponse {
  status: "fail";
  statusCode: number;
  errors: error | error[];
}

export interface SuccessResponse<T> {
  statusCode: number;
  data: T;
}